Move usePopup.ts file to same folder as popup.ts for readability
